+++ Source/UnrealCAPI/Public/uec_api.h
@@ -22,7 +22,7 @@
 #endif
 
 #define UEC_ABI_MAJOR 1u
-#define UEC_ABI_MINOR 25u
+#define UEC_ABI_MINOR 26u
 
 #ifdef __cplusplus
 extern "C" {
@@ -75,7 +75,8 @@ enum {
     UEC_CAPABILITY_CAMERA = UINT64_C(1) << 18,
     UEC_CAPABILITY_SAVE_DATA = UINT64_C(1) << 19,
     UEC_CAPABILITY_THREADING = UINT64_C(1) << 20,
-    UEC_CAPABILITY_MOVEMENT = UINT64_C(1) << 21
+    UEC_CAPABILITY_MOVEMENT = UINT64_C(1) << 21,
+    UEC_CAPABILITY_PRESENTATION = UINT64_C(1) << 22
 };
 
 typedef struct uec_context uec_context;
@@ -407,6 +408,11 @@ typedef struct uec_api {
                                                    uec_bool force);
     uec_result (UEC_CALL *jump_character)(uec_actor* character);
     uec_result (UEC_CALL *stop_character_jumping)(uec_actor* character);
+    uec_result (UEC_CALL *set_static_mesh)(uec_scene_component* component,
+                                           uec_object* mesh);
+    uec_result (UEC_CALL *set_skeletal_mesh)(uec_scene_component* component,
+                                             uec_object* mesh,
+                                             uec_bool reinitialize_pose);
 } uec_api;
 
 /* Bootstrap entry point. The returned function table remains valid until the